Remove `cli-main.js`
