+++ package.json
@@ -79,7 +79,6 @@
 		"fs-extra": "^10.0.0",
 		"get-stdin": "^8.0.0",
 		"globby": "^9.2.0",
-		"has-flag": "^4.0.0",
 		"imurmurhash": "^0.1.4",
 		"is-path-inside": "^3.0.3",
 		"json-stable-stringify-without-jsonify": "^1.0.1",
@@ -93,7 +92,6 @@
 		"p-reduce": "^2.1.0",
 		"path-exists": "^4.0.0",
 		"prettier": "^2.2.1",
-		"resolve-cwd": "^3.0.0",
 		"semver": "^7.3.5",
 		"slash": "^3.0.0",
 		"to-absolute-glob": "^2.0.2",
