+++ test/print-config.js
@@ -8,7 +8,7 @@ import xo from '../index.js';
 const {__dirname} = createEsmUtils(import.meta);
 process.chdir(__dirname);
 
-const main = (arguments_, options) => execa(path.join(__dirname, '../cli-main.js'), arguments_, options);
+const main = (arguments_, options) => execa(path.join(__dirname, '../cli.js'), arguments_, options);
 
 const hasUnicornPlugin = config => config.plugins.includes('unicorn');
 const hasPrintConfigGlobal = config => Object.keys(config.globals).includes('printConfig');
